pyproject.toml
[project]
name = "finance-news-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "openai>=1.77.0",
    "pdfkit>=1.0.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
